+++ src/features/store.ts
@@ -64,6 +64,7 @@ import { apiSlice as serviceAdminBoardApiSlice } from './adminBoard/serviceAdmin
 import { apiSlice as companyRoleApiSlice } from './companyRoles/companyRoleApiSlice'
 import { apiSlice as certificationApiSlice } from './certification/certificationApiSlice'
 import { apiSlice as userManagementApiSlice } from './appManagement/userManagementApiSlice'
+import { apiSlice as companyWalletApiSlice } from './compayWallet/companyWalletApiSlice'
 import { apiSlice as deleteCompanyApiSlice } from './deleteCompany/deleteCompanyApiSlice'
 import { apiSlice as registrationApiSlice } from './registration/registrationApiSlice'
 import { apiSlice as staticContentApiSlice } from './staticContent/staticContentApiSlice'
@@ -120,6 +121,7 @@ export const reducers = {
   [certificationApiSlice.reducerPath]: certificationApiSlice.reducer,
   [userManagementApiSlice.reducerPath]: userManagementApiSlice.reducer,
   [usecaseApiSlice.reducerPath]: usecaseApiSlice.reducer,
+  [companyWalletApiSlice.reducerPath]: companyWalletApiSlice.reducer,
   [deleteCompanyApiSlice.reducerPath]: deleteCompanyApiSlice.reducer,
   [registrationApiSlice.reducerPath]: registrationApiSlice.reducer,
   [staticContentApiSlice.reducerPath]: staticContentApiSlice.reducer,
@@ -154,6 +156,7 @@ export const store = configureStore({
       .concat(certificationApiSlice.middleware)
       .concat(userManagementApiSlice.middleware)
       .concat(usecaseApiSlice.middleware)
+      .concat(companyWalletApiSlice.middleware)
       .concat(registrationApiSlice.middleware)
       .concat(staticContentApiSlice.middleware)
       .concat(deleteCompanyApiSlice.middleware)